pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "tradingagents"
version = "0.3.0"
description = "TradingAgents: Multi-Agents LLM Financial Trading Framework"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28",
    "pandas>=2.3.0",
    "numpy>=1.24.0",
    "pytz>=2025.2",
    "python-dateutil>=2.9.0",
    "questionary>=2.1.0",
    "rank-bm25>=0.2.2",
    "requests>=2.32.4",
    "rich>=14.0.0",
    "typer>=0.21.0",
    "setuptools>=80.9.0",
    "stockstats>=0.6.5",
    "typing-extensions>=4.14.0",
    "yfinance>=0.2.63",
    "python-dotenv>=1.0.0",
]

[project.scripts]
tradingagents = "cli.app:app"

[tool.setuptools.packages.find]
include = ["tradingagents*", "cli*"]
